CHL-cambios de reunion

+++ CRS_DOC/TAB/ctb002.sql
@@ -7,11 +7,11 @@ FECHA:		03-12-2017
 	
 CREATE TABLE ctb002 
 (
-va_cod_cap		TINYINT			NOT NULL,	--Capitulo/Agrupador (1 Número)
+va_cod_cap		INT				NOT NULL,	--Capitulo/Agrupador (1 Número)
 va_nom_cap		VARCHAR(20)		NOT NULL,	--Nombre Capitulo/Agrupador
 va_tip_cap		CHAR(01)		NOT NULL,	--Tipo de capitulo (A=activo; P=pasivo; T=patrimonio; I=ingreso; E=egreso; C=costo)
 va_tra_cap		CHAR(1)			NOT NULL,	--Tratamiento (D=Deudor ; A=Acreedor)
-va_cen_cto		TINYINT			NOT NULL,	--Usa Centro de Costo(0=No ; 1=Si Usa)
+va_cen_cto		INT				NOT NULL,	--Usa Centro de Costo(0=No ; 1=Si Usa)
 va_est_ado		CHAR(1)			NOT NULL	--Estado (H=Habilitado ; N=Deshabilitado)		  
       
 CONSTRAINT pk1_ctb002 PRIMARY KEY(va_cod_cap)
